Moving python using code from circuit_reader to py_methods
